# docs/swagger.yaml
basePath: /
definitions:
  github_com_pborgen_future-api_internal_handler_httputil.ErrorResponse:
    properties:
      error:
        example: invalid trainer_id
        type: string
    type: object
  github_com_pborgen_future-api_internal_model.Appointment:
    properties:
      ends_at:
        example: "2026-04-06T09:30:00-07:00"
        type: string
      id:
        example: 42
        type: integer
      starts_at:
        example: "2026-04-06T09:00:00-07:00"
        type: string
      trainer_id:
        example: 1
        type: integer
      user_id:
        example: 2
        type: integer
    type: object
  github_com_pborgen_future-api_internal_model.CreateAppointmentRequest:
    properties:
      ends_at:
        example: "2026-04-06T09:30:00-07:00"
        type: string
      starts_at:
        example: "2026-04-06T09:00:00-07:00"
        type: string
      trainer_id:
        example: 1
        type: integer
      user_id:
        example: 2
        type: integer
    type: object
  github_com_pborgen_future-api_internal_model.Slot:
    properties:
      ends_at:
        example: "2026-04-06T09:30:00-07:00"
        type: string
      starts_at:
        example: "2026-04-06T09:00:00-07:00"
        type: string
    type: object
info:
  contact: {}
  description: HTTP/JSON API for scheduling 30-minute trainer appointments. All times
    are RFC3339; business hours are evaluated in America/Los_Angeles (M-F 8am-5pm).
  title: Future Appointment API
  version: "1.0"
paths:
  /appointments:
    post:
      consumes:
      - application/json
      description: Creates a 30-minute appointment for the given trainer and user.
        The slot must start on :00 or :30, fall within M-F 8am-5pm Pacific, and not
        overlap an existing booking for the trainer.
      parameters:
      - description: Appointment to create
        in: body
        name: appointment
        required: true
        schema:
          $ref: '#/definitions/github_com_pborgen_future-api_internal_model.CreateAppointmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_pborgen_future-api_internal_model.Appointment'
        "400":
          description: malformed request
          schema:
            $ref: '#/definitions/github_com_pborgen_future-api_internal_handler_httputil.ErrorResponse'
        "409":
          description: trainer already booked at that time
          schema:
            $ref: '#/definitions/github_com_pborgen_future-api_internal_handler_httputil.ErrorResponse'
        "422":
          description: outside business hours, wrong duration, or not on :00/:30
          schema:
            $ref: '#/definitions/github_com_pborgen_future-api_internal_handler_httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_pborgen_future-api_internal_handler_httputil.ErrorResponse'
      summary: Book a new appointment
      tags:
      - appointments
  /trainers/{trainer_id}/appointments:
    get:
      description: Returns every appointment booked with the given trainer, ordered
        by start time.
      parameters:
      - description: Trainer ID
        in: path
        name: trainer_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_pborgen_future-api_internal_model.Appointment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pborgen_future-api_internal_handler_httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_pborgen_future-api_internal_handler_httputil.ErrorResponse'
      summary: List a trainer's scheduled appointments
      tags:
      - appointments
  /trainers/{trainer_id}/availability:
    get:
      description: Returns every bookable 30-minute slot for the trainer between starts_at
        and ends_at. Slots that overlap an existing booking are excluded. Times must
        be RFC3339.
      parameters:
      - description: Trainer ID
        in: path
        name: trainer_id
        required: true
        type: integer
      - description: Window start (RFC3339)
        example: "2026-04-06T08:00:00-07:00"
        in: query
        name: starts_at
        required: true
        type: string
      - description: Window end (RFC3339)
        example: "2026-04-06T17:00:00-07:00"
        in: query
        name: ends_at
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_pborgen_future-api_internal_model.Slot'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_pborgen_future-api_internal_handler_httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_pborgen_future-api_internal_handler_httputil.ErrorResponse'
      summary: List available appointment slots for a trainer
      tags:
      - appointments
swagger: "2.0"